Prefabs, Particles, and Interactions
Made fog, rain, and fireflies.
Coded cache interactions (except for buttons).
Coded porch light and light's interaction with trash.
Cleared out original test art (we should clear out unused and test art
when lvl 1 is finished and we leave the sandbox. It'll cut down on our
file size a LOT)
Changed the names of a few things
Created Prefabs out of as much of the sandbox as possible for switching
them over to Lvl 1